feat(modal-pages): add delay before calling onClose to let animations finish
