+++ packages/application-components/src/components/modal-pages/tabular-modal-page/README.md
@@ -49,6 +49,7 @@ Tabular Modal pages are controlled components used to render a page with custom 
 | `isPrimaryButtonDisabled`       | `boolean`                  |    -     | -                        | false                                                                                                                                                                                                                                                                                                       | Indicates whether primary button is disabled or not |
 | `dataAttributesSecondaryButton` | `object`                   |    -     | -                        | -                                                                                                                                                                                                                                                                                                           | Use this prop to pass `data-` attributes to the secondary button |
 | `dataAttributesPrimaryButton`   | `object`                   |    -     | -                        | -                                                                                                                                                                                                                                                                                                           | Use this prop to pass `data-` attributes to the primary button |
+| `shouldDelayOnClose`            | `bool`                     |    -     | `true`                   | Sets wether the ModalPage should delay calling its onClose function to allow the closing animation to finish. This can be turned off if the developer is controlling the ModalPage only through the `isOpen` prop, and not abruptly mounting/unmounting it or one of its parent elements. |
 
 ## Static properties
